Alert Dialog and UI change for deleting posts and comments

+++ app/src/main/java/com/example/seniorproject/MainForum/Posts/CommunityPosts.kt
@@ -1,8 +1,14 @@
 package com.example.seniorproject.MainForum.Posts
 
+import android.content.DialogInterface
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
+import android.graphics.drawable.Drawable
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
@@ -23,6 +29,9 @@ import javax.inject.Inject
 class CommunityPosts : AppCompatActivity() {
 
     private lateinit var adapter: CustomAdapter
+
+    var swipeBackground: ColorDrawable = ColorDrawable(Color.parseColor("#FF0000"))
+    lateinit var deleteIcon: Drawable
 
     @Inject
     lateinit var factory: ViewModelProvider.Factory
@@ -62,18 +71,85 @@ class CommunityPosts : AppCompatActivity() {
 
         }
 
+        deleteIcon = ContextCompat.getDrawable(this, R.drawable.ic_delete_24px)!!
+
+
         val itemTouchHelperCallback =
             object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
                 override fun onSwiped(viewHolders: RecyclerView.ViewHolder, position: Int) {
                     val postkey: String? =
                         adapter.removeItem(viewHolders as CustomViewHolders, position)
 
+
+                    var builder = AlertDialog.Builder(this@CommunityPosts, R.style.AppTheme_AlertDialog)
+
                     //.getStringExtra("Classkey")
                     //val postkey = intent.getStringExtra("author")
-                    myViewModel.deletePost(postkey!!, className)
+                    //myViewModel.deletePost(postkey!!, className)
                     //myViewModel.deletePost()
+                    builder.setTitle("Are you sure?")
+                    builder.setMessage("You cannot restore posts that have been deleted.")
+                    builder.setPositiveButton("DELETE",
+                        { dialogInterface: DialogInterface?, i: Int ->
+                            myViewModel.deletePost(postkey!!, className)
+                        })
+                    builder.setNegativeButton("CANCEL",
+                        { dialogInterface: DialogInterface?, i: Int ->
+                            builder.setCancelable(true)
+                        })
 
+                    val msgdialog: AlertDialog = builder.create()
+
+                    msgdialog.show()
+
+                    //myViewModel.deletePost(postkey!!, className)
+                    //classes_post_RV.removeViewAt(viewHolders.adapterPosition)
+                    adapter.notifyItemRemoved(viewHolders.adapterPosition)
+                    adapter.notifyItemRangeChanged(viewHolders.adapterPosition, adapter.itemCount)
+                    classes_post_RV.adapter = adapter
+
+                    //adapter.notifyDataSetChanged()
+                    //adapter.notifyItemRangeRemoved(viewHolders.adapterPosition, 1)
+                    /*adapter.notifyItemRangeChanged(viewHolders.adapterPosition, adapter.itemCount - viewHolders.adapterPosition+1)*/
+                    //adapter.notifyDataSetChanged()
+
+                    // adapter.notifyItemRangeRemoved(viewHolders.adapterPosition, 1)
+                    // adapter.notifyItemRemoved(viewHolders.adapterPosition)
+                    // adapter.notifyItemRangeChanged(viewHolders.adapterPosition, )
                     //  obse.onChanged(pos)
+
+                }
+
+                override fun onChildDraw(
+                    c: Canvas,
+                    recyclerView: RecyclerView,
+                    viewHolder: RecyclerView.ViewHolder,
+                    dX: Float,
+                    dY: Float,
+                    actionState: Int,
+                    isCurrentlyActive: Boolean
+                ) {
+                    val itemView = viewHolder.itemView
+                    val iconMargin = (itemView.height - deleteIcon.intrinsicHeight)/2
+
+                    if(dX < 0) {
+                        swipeBackground.setBounds(itemView.right + dX.toInt(), itemView.top, itemView.right, itemView.bottom)
+                        deleteIcon.setBounds(itemView.right - iconMargin - deleteIcon.intrinsicWidth, itemView.top + iconMargin,
+                            itemView.right - iconMargin, itemView.bottom - iconMargin)
+                    }
+
+                    swipeBackground.draw(c)
+                    deleteIcon.draw(c)
+
+                    super.onChildDraw(
+                        c,
+                        recyclerView,
+                        viewHolder,
+                        dX,
+                        dY,
+                        actionState,
+                        isCurrentlyActive
+                    )
                 }
 
                 override fun onMove(
